// types/api/cliente.ts
// Interface automatically generated by schemas-to-ts

import { Venda } from './venda';
import { Venda_Plain } from './venda';
import { AdminPanelRelationPropertyModification } from '../common/AdminPanelRelationPropertyModification';

export interface Cliente {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    _softDeletedAt?: Date;
    _softDeletedById?: number;
    _softDeletedByType?: string;
    nome: string;
    cpf?: string;
    celular?: string;
    insta?: string;
    cep?: string;
    endereco?: string;
    distrito?: string;
    cidade?: string;
    estado?: string;
    numero?: string;
    complemento?: string;
    email?: string;
    vendas: { data: Venda[] };
    id_asaas?: string;
  };
}
export interface Cliente_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  _softDeletedAt?: Date;
  _softDeletedById?: number;
  _softDeletedByType?: string;
  nome: string;
  cpf?: string;
  celular?: string;
  insta?: string;
  cep?: string;
  endereco?: string;
  distrito?: string;
  cidade?: string;
  estado?: string;
  numero?: string;
  complemento?: string;
  email?: string;
  vendas: Venda_Plain[];
  id_asaas?: string;
}

export interface Cliente_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  _softDeletedAt?: Date;
  _softDeletedById?: number;
  _softDeletedByType?: string;
  nome: string;
  cpf?: string;
  celular?: string;
  insta?: string;
  cep?: string;
  endereco?: string;
  distrito?: string;
  cidade?: string;
  estado?: string;
  numero?: string;
  complemento?: string;
  email?: string;
  vendas: number[];
  id_asaas?: string;
}

export interface Cliente_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  _softDeletedAt?: Date;
  _softDeletedById?: number;
  _softDeletedByType?: string;
  nome: string;
  cpf?: string;
  celular?: string;
  insta?: string;
  cep?: string;
  endereco?: string;
  distrito?: string;
  cidade?: string;
  estado?: string;
  numero?: string;
  complemento?: string;
  email?: string;
  vendas: AdminPanelRelationPropertyModification<Venda_Plain>;
  id_asaas?: string;
}